Restore load YouTube player on new
Only after the onReady event is fired is the YouTube player ready to
load videos. Since this is the actual true signal of readiness we update
the initPlayer promise in YoutubePlayer to resolve in that event.

+++ app/javascript/controllers/player_controller.js
@@ -267,7 +267,7 @@ export default class extends Controller {
         // to be available when video metadata is loaded, which happens on first
         // play only.
         if(this.hasDurationTarget && this.durationTarget.value !== "") {
-          this.durationTarget.value = parseFloat(this.player.duration)
+          this.durationTarget.value = parseInt(this.player.duration)
         }
       }
     }
